Create Spanish-419.txt
I'm not sure about the translation of "gamebox" into Spanish so I just picked one that might fit but I still think a better can be proposed.
Also, I don't know if with Watchdog you refer to the executable name (Watchdog.exe) or if I should translate the word watchdog literally. As it was in uppercase in English I supposed it is referring to the executable name. And lastly, I didn't know how to name this localization file, it has been made with Spanish-Latin America in mind but there is no word that must be translated differently for Spanish-Spain so this can be used for both.